+++ libft/srcs/lst/ft_lstsize.c
@@ -6,7 +6,7 @@
 /*   By: abidolet <abidolet@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/20 20:18:35 by abidolet          #+#    #+#             */
-/*   Updated: 2025/07/04 23:29:45 by abidolet         ###   ########.fr       */
+/*   Updated: 2025/07/06 14:53:33 by abidolet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_lstsize(t_list *lst)
 {
-	unsigned long	res;
+	unsigned long long	res;
 
 	res = 0;
 	while (lst)
