+++ libft/srcs/mem/ft_bzero.c
@@ -6,11 +6,13 @@
 /*   By: abidolet <abidolet@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/14 17:13:09 by abidolet          #+#    #+#             */
-/*   Updated: 2025/07/04 23:29:56 by abidolet         ###   ########.fr       */
+/*   Updated: 2025/07/06 15:15:11 by abidolet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	ft_bzero(void *s, unsigned long n)
+#include <stddef.h>
+
+void	ft_bzero(void *s, size_t n)
 {
 	while (n-- > 0)
 		((unsigned char *)s)[n] = 0;
