+++ libft/srcs/mem/ft_memcmp.c
@@ -10,7 +10,9 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_memcmp(const void *s1, const void *s2, unsigned long n)
+#include <stddef.h>
+
+int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
 	while (n-- > 0)
 		if (*((unsigned char *)s1++) != *((unsigned char *)s2++))
