+++ libft/srcs/mem/ft_memcpy.c
@@ -6,11 +6,13 @@
 /*   By: abidolet <abidolet@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/20 20:18:55 by abidolet          #+#    #+#             */
-/*   Updated: 2025/07/04 23:30:04 by abidolet         ###   ########.fr       */
+/*   Updated: 2025/07/06 15:16:19 by abidolet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memcpy(void *dest, const void *src, unsigned long n)
+#include <stddef.h>
+
+void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
 	if (!dest && !src)
 		return (dest);
