+++ libft/srcs/mem/ft_memset.c
@@ -6,11 +6,13 @@
 /*   By: abidolet <abidolet@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/20 20:19:16 by abidolet          #+#    #+#             */
-/*   Updated: 2025/07/04 23:30:08 by abidolet         ###   ########.fr       */
+/*   Updated: 2025/07/06 15:18:04 by abidolet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memset(void *s, int c, unsigned long n)
+#include <stddef.h>
+
+void	*ft_memset(void *s, int c, size_t n)
 {
 	while (n-- > 0)
 		*(unsigned char *)(s + n) = (unsigned char)c;
