+++ libft/srcs/mem/ft_realloc.c
@@ -6,14 +6,14 @@
 /*   By: abidolet <abidolet@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/31 20:06:52 by abidolet          #+#    #+#             */
-/*   Updated: 2025/07/04 23:22:10 by abidolet         ###   ########.fr       */
+/*   Updated: 2025/07/06 14:42:36 by abidolet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft/mem.h"
 #include <stdlib.h>
 
-void	*realloc(void *ptr, unsigned long size)
+void	*realloc(void *ptr, size_t size)
 {
 	void	*res;
