+++ libft/srcs/str/ft_strcpy.c
@@ -6,7 +6,7 @@
 /*   By: abidolet <abidolet@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/19 22:58:53 by abidolet          #+#    #+#             */
-/*   Updated: 2025/07/04 23:22:10 by abidolet         ###   ########.fr       */
+/*   Updated: 2025/07/06 15:13:42 by abidolet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strcpy(char *dest, const char *src)
 {
-	unsigned long	size_src;
+	size_t	size_src;
 
 	size_src = ft_strlen(src) + 1;
 	while (size_src--)
