+++ libft/srcs/str/ft_strlcpy.c
@@ -13,7 +13,7 @@
 #include "libft/str.h"
 #include <stddef.h>
 
-size_t	ft_strlcpy(char *dest, const char *src, unsigned long size)
+size_t	ft_strlcpy(char *dest, const char *src, size_t size)
 {
 	if (!dest || !src)
 		return (ft_strlen(src));
