+++ libft/srcs/str/ft_strncmp.c
@@ -10,7 +10,9 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strncmp(const char *s1, const char *s2, unsigned long n)
+#include <stddef.h>
+
+int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
 	while (n-- && (*s1 || *s2) && *s1++ == *s2++)
 		;
